+++ crates/zilog-z80/src/cpu.rs
@@ -1030,8 +1030,8 @@ impl Cpu for Z80 {
         }
     }
 
-    fn pc(&self) -> u16 {
-        self.regs.pc
+    fn pc(&self) -> u32 {
+        u32::from(self.regs.pc)
     }
 
     fn registers(&self) -> Self::Registers {
